% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.folds}
\alias{cv.folds}
\title{cv.folds}
\usage{
cv.folds(n, folds = 10)
}
\arguments{
\item{n}{number of observations to split}

\item{folds}{number of folds to create}
}
\value{
a list of length 5, where each item in the list is a vector
of test set indices.
}
\description{
split n data points into a specific number of folds.
This function was copied over from CRAN/scout
}
