% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\name{est_cov}
\alias{est_cov}
\title{est_cov}
\usage{
est_cov(
  X,
  Y = NULL,
  method = c("default", "winsor", "wrap", "ddc"),
  correlation = FALSE
)
}
\arguments{
\item{X}{(n x p) predictor matrix}

\item{Y}{(n x 1) response vector or NULL}

\item{method}{string indicating which covariance estimator
to compute. One of "default", "winsor", "wrap", "ddc".}

\item{correlation}{if TRUE, returns a correlation matrix}
}
\value{
a (p x p) covariance matrix or length-p vector
}
\description{
Computes a covariance estimate based on the type
specificed by the user. This function assumes that the data
has already been standardized.
}
