% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso_select.R
\name{glasso_cv}
\alias{glasso_cv}
\title{\code{glasso_cv}}
\usage{
glasso_cv(
  X,
  K,
  standardize,
  centerFun,
  scaleFun,
  cov_method,
  crit,
  folds = NULL,
  nlambda = 100,
  lambda.min.ratio = 0.1,
  lambdas = NULL,
  scr = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{X}{feature matrix, \eqn{n \times p}}

\item{standardize}{whether or not to standardize the data before running glasso}

\item{centerFun}{a function to compute an estiamte of the center of a variable.
This is ignored if standardize is \code{FALSE}}

\item{scaleFun}{a function to compute an estimate of the scale of a variable.
This is ignored if standardize is \code{FALSE}}

\item{cov_method}{a string indicating which covariance matrix to use. One of "default" or "winsor"}

\item{crit}{criteria to select the optimal lambda. one of "bic" or "loglik"}

\item{folds}{a list of length K containing the indices of the test set in each CV fold}

\item{nlambda}{number of lambdas to optimize over}

\item{lambda.min.ratio}{smallest value of lambda as a fraction of lambda_max}

\item{lambdas}{sequence of lambda values. This function will generate its own lambda sequence
if this is NULL. If provided, \code{nlambda} and \code{lambda.min.ratio} are ignored.}

\item{scr}{whether to use lossy screening in huge.glasso}

\item{verbose}{whether to let huge.glasso print progress messages}
}
\value{
result of running huge::huge.glasso on the ideal lambda found from cv
}
\description{
Performs K-fold cross-validation to select the best graphical model using
huge::huge.glasso
}
