% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso_select.R
\name{huge_glasso_lambda_seq}
\alias{huge_glasso_lambda_seq}
\title{\code{huge_glasso_lambda_seq}}
\usage{
huge_glasso_lambda_seq(S, nlambda, lambda.min.ratio = 0.1)
}
\arguments{
\item{S}{covariance matrix estimate cov(X)}

\item{nlambda}{number of lambdas to return}

\item{lambda.min.ratio}{smallest value of lambda as a fraction of lambda_max}
}
\value{
numeric vector of length nlambda, in decreasing order, of log-spaced
lambda values
}
\description{
Computes the lambda sequence for huge::huge.glasso
}
