% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robscout.R
\name{robscout}
\alias{robscout}
\title{robscout}
\usage{
robscout(
  X,
  Y,
  est_type = c("glasso", "gridge"),
  lam1,
  lam2,
  centerFun = stats::median,
  scaleFun = stats::mad
)
}
\description{
robscout
}
