% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scout_stepwise.R
\name{scout_1something_stepwise}
\alias{scout_1something_stepwise}
\title{Scout(1,.) Stepwise}
\usage{
scout_1something_stepwise(
  X,
  Y,
  p2,
  K = 5,
  nlambda1 = 100,
  lambda1.min.ratio = 0.1,
  nlambda2 = 100,
  lambda2.min.ratio = 0.1,
  cov_method = "default",
  glasso_crit = "bic",
  ddc_first = FALSE,
  ddc_with_response = FALSE,
  standardize = TRUE,
  centerFun = mean,
  scaleFun = sd,
  rescale = TRUE
)
}
\arguments{
\item{X}{n x p data matrix}

\item{Y}{n x 1 response vector}

\item{p2}{regularization type in second step. one of NULL or 1}

\item{K}{number of folds to use in CV, only applicable for p2 = 1}

\item{nlambda1}{number of regularization params to use in glasso step}

\item{lambda1.min.ratio}{smallest value of lambda1 as a fraction of lambda1_max}

\item{nlambda2}{number of regularization params to use in lasso step}

\item{lambda2.min.ratio}{smallest value of lambda2 as a fraction of lambda2_max}

\item{cov_method}{a string indicating which covariance matrix to use. See below for details}

\item{glasso_crit}{a string indicating which criteria to minimize in glasso}

\item{ddc_first}{whether or not to detect and impute cellwise outliers using
DDC before computing estimates}

\item{ddc_with_response}{whether or not to concatenate Y before running DDC. Only
applicable if \code{ddc_first} is \code{TRUE}}

\item{standardize}{whether or not to standardize variables before
computing estiamtes}

\item{centerFun}{a function to compute an estiamte of the center of a variable.
This is ignored if standardize is \code{FALSE}}

\item{scaleFun}{a function to compute an estimate of the scale of a variable.
This is ignored if standardize is \code{FALSE}}

\item{rescale}{should coefficients be re-scaled by a constant?}
}
\value{
list of
\itemize{
\item \code{g.res}: result of graphical lasso step
\item \code{cv.res}: result of cross-validated lasso step, or NA if p2 = NULL
\item \code{mod}: final scout model using cross-validated parameters
}
}
\description{
Computes the best inverse covariance estimate using the glasso and uses that
in the second step of the Scout.
}
\details{
\code{cov_method} can be one of the following:
\itemize{
\item \code{default}: computes the default covariance with \code{\link{cov}}
\item \code{winsor}: computes the covariance matrix based on adjusted multivariate Winsorization,
as seen in Lafit et al. 2022.
}
}
