% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\name{cov_winsor}
\alias{cov_winsor}
\title{cov_winsor}
\usage{
cov_winsor(X, Y = NULL)
}
\arguments{
\item{X}{a nxp matrix}

\item{Y}{a nx1 vector or NULL}
}
\value{
a pxp matrix if Y is NULL; otherwise, a length-p vector.
}
\description{
computes an estimate of cov(X,Y) with adjusted
multivariate Winsorization, as described in Lafit et al. 2022.
This function will standardize the data using median and mad.
Hence, if the data has already been standardized with those
robust center and scale functions, standardization here
will have no effect.
}
