% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso_select.R
\name{glasso_select}
\alias{glasso_select}
\title{\code{glasso_select}}
\usage{
glasso_select(
  X,
  nlambda = 100,
  lambda_min_ratio = 0.1,
  crit = "bic",
  standardize = TRUE,
  scr = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{X}{feature matrix, \eqn{n \times p}}

\item{nlambda}{number of lambdas to optimize over}

\item{lambda_min_ratio}{smallest value of lambda as a fraction of lambda_max}

\item{crit}{criteria to select the optimal lambda. one of "bic" or "loglik"}

\item{scr}{whether to use lossy screening in huge.glasso}

\item{verbose}{whether to let huge.glasso print progress messages}
}
\value{
list of:
\itemize{
\item \code{icov}: matrix - inverse covariance estimate based on best lambda
\item \code{best_lambda}: numeric - best lambda selected based on \code{crit}
\item \code{lambda}: numeric vector - sequence of lambdas that was used for selection
\item \code{errors}: numeric vector - \code{crit} values corresponding to each value in
\code{lambda}
}
}
\description{
Uses huge::huge.glasso to estimate an inverse covariance matrix given
a data matrix.
}
