% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robscout.R
\name{huge_glasso_lambda_seq}
\alias{huge_glasso_lambda_seq}
\title{\code{huge_glasso_lambda_seq}}
\usage{
huge_glasso_lambda_seq(cov_x, nlambda, lambda_min_ratio)
}
\arguments{
\item{cov_x}{covariance matrix estimate}

\item{nlambda}{number of lambdas to return}

\item{lambda_min_ratio}{smallest value of lambda as a fraction of lambda_max}
}
\value{
numeric vector of length nlambda, in decreasing order, of log-spaced
lambda values
}
\description{
Computes the lambda sequence for huge::huge.glasso
}
